Update tf_projectile_rocket.cpp
Added support for team skins with rockets. The method CTFProjectile_Rocket::Deflected calls for switching team skins, but this doesn't occur when the rocket is spawned. As a result, the rocket will always have the RED team's skin, never changing skin unless airblasted.

This commit fixes that, by setting the team skin when spawned.